handle login form and break it down in subparts
