+++ web/src/components/Auth/EmailPassword/EmailPasswordSignupForm.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function EmailPasswordSignupForm() {
+  return <div>EmailPasswordSignupForm</div>;
+}
+
+export default EmailPasswordSignupForm;
