+++ web/src/components/Auth/SocialSignIn/oauthProvidersButtons.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import OctaSignIn from "./OctaSignIn.tsx";
+import { AuthProviders } from "../utils/AuthProviders.ts";
 
 export const oAuthProviders = {
-  OKTA: <OctaSignIn />,
+  [AuthProviders.OKTA]: <OctaSignIn />,
 };
