+++ web/src/components/Auth/utils/AuthProviders.ts
@@ -1,0 +1,4 @@
+export const AuthProviders = {
+  EMAIL: "EMAIL",
+  OKTA: "OKTA",
+};
